Title: 최소 에너지 비용, Time: 288ms, Memory: 23MB, Status: Passed - CodeTree

+++ 250314/최소 에너지 비용/minimum-energy-cost.java
@@ -3,13 +3,13 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
-        int[] dist = new int[n + 1];
+        long[] dist = new long[n + 1];
         int[] cost = new int[n + 1];
         for (int i = 0; i < n-1; i++)
             dist[i] = sc.nextInt();
         for (int i = 0; i < n; i++)
             cost[i] = sc.nextInt();
-        int min_cost = cost[0];
+        long min_cost = cost[0];
         long result = 0;
         for(int i=0;i<n-1;i++){
             min_cost = Math.min(min_cost,cost[i]);
